src/test/java/com/leetcode/questions/FizzBuzzTest.java
package com.leetcode.questions;

import org.junit.Before;
import org.junit.Test;

import static org.assertj.core.api.Assertions.assertThat;

public class FizzBuzzTest {
    private FizzBuzz fizzBuzz;

    @Before
    public void setUp() {
        fizzBuzz = new FizzBuzz();
    }

    @Test
    public void test_1() {
        assertThat(fizzBuzz.fizzBuzz(1)).containsOnly("1");
    }

    @Test
    public void test_2() {
        assertThat(fizzBuzz.fizzBuzz(15)).containsExactly("1",
                "2",
                "Fizz",
                "4",
                "Buzz",
                "Fizz",
                "7",
                "8",
                "Fizz",
                "Buzz",
                "11",
                "Fizz",
                "13",
                "14",
                "FizzBuzz");
    }
}